Upgrade powsybl 6.0.0 (#139)
* first commit upgrade powsybl

* upgrade powsybl and refactoring of export cgmes

* upgrade farao dichotomye=

* Adding unit tests

* Update dependencies

* Add business logs to stdout and catch exception

* Update rao-runner and fix unit test

* Review remarks

* upgrade dependencies

* refactoring CgmesExport for SSH

* Delete MergedNetworkData

* refactoring and adding test export ssh with error

* Last commit for reviewer

+++ gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/dichotomy/shift/ZonalScalableProvider.java
@@ -50,7 +50,7 @@ public class ZonalScalableProvider {
 
     private Scalable getCountryGeneratorsScalableForFR(Network network) {
         List<Scalable> scalables = new ArrayList<>();
-        List<Float> percentages = new ArrayList<>();
+        List<Double> percentages = new ArrayList<>();
         List<Generator> generators;
         generators = network.getGeneratorStream()
                 .filter(generator -> Country.FR.equals(generator.getTerminal().getVoltageLevel().getSubstation().map(Substation::getNullableCountry).orElse(null)))
@@ -60,7 +60,7 @@ public class ZonalScalableProvider {
         double totalCountryP = generators.stream().mapToDouble(NetworkUtil::pseudoTargetP).sum();
         //calculate factor of each generator
         generators.forEach(generator -> {
-            float generatorPercentage = (float) (100 * NetworkUtil.pseudoTargetP(generator) / totalCountryP);
+            double generatorPercentage =  100 * NetworkUtil.pseudoTargetP(generator) / totalCountryP;
             percentages.add(generatorPercentage);
             scalables.add(Scalable.onGenerator(generator.getId()));
         });
